thanks to davorin kunstelj for spotting that numbers between -1 and 0 were not printfd properly
git-svn-id: https://pyjamas.svn.sourceforge.net/svnroot/pyjamas/trunk@1031 7a2bd370-bda8-463c-979e-2900ccfb811e

+++ pyjs/src/pyjs/builtin/public/sprintf.js
@@ -117,7 +117,7 @@ sprintfWrapper = {
 				min: match[6] || 0,
 				precision: match[8],
 				code: match[9] || '%',
-				negative: parseInt(param) < 0 ? true : false,
+				negative: parseFloat(param) < 0 ? true : false,
 				argument: String(param)
 			};
 		}
